+++ exemple/05_config_enable_tool.php
@@ -16,10 +16,10 @@ class Foo
     public function __construct()
     {
         // Disable tool manual
-        Config::set(Config::DISABLE_TOOL, true);
+        Config::set(Config::ENABLE_TOOL, false);
         // OR use a string
         // Disable tool by the ENV by using the APP_DEBUG value.
-        Config::set(Config::DISABLE_TOOL, 'ENV:APP_DEBUG');
+        Config::set(Config::ENABLE_TOOL, 'ENV:APP_DEBUG');
 
         $this->taskA();
         $this->taskB();